Add files via upload
Create a custom list view on account object. The list view should filter on the basis of email and last name. The default list view should show all account records. The search functionality should be dynamic not on the click of search button. If user tries to search for email id for eg test@gmail.com and then removes the search text, then the list should show all account records.